% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abw.R
\name{abw}
\alias{abw}
\title{Calculate Adjusted Body Weight}
\usage{
abw(height, weight, female = TRUE, formula = "Broca")
}
\arguments{
\item{height}{Height in meters.}

\item{weight}{Weight in kilograms.}

\item{female}{Boolean indicating female sex.}

\item{formula}{Formula to use for calculating ideal body weight, see documentation for \code{\link[=ibw]{ibw()}} for options}
}
\value{
A number of kilograms.
}
\description{
Calculates adjusted body weight.
}
\examples{
abw(1.81, 71.2, female = FALSE, formula = 'Hawmi')
}
\seealso{
\code{\link[=ibw]{ibw()}}
}
