% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_category.R
\name{bmi_category}
\alias{bmi_category}
\title{Print BMI Category}
\usage{
bmi_category(bmi)
}
\arguments{
\item{bmi}{A body mass index.}
}
\value{
A string containing the body mass index’s weight category.
}
\description{
Prints the \href{https://www.who.int/europe/news-room/fact-sheets/item/a-healthy-lifestyle---who-recommendations}{weight category according to the World Health Organization}
for a given body mass index.
}
\examples{
bmi_category(21.7)
}
