% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_limits.R
\name{bmi_limits}
\alias{bmi_limits}
\title{Calculate Body Mass Index Weight Limits}
\usage{
bmi_limits(height)
}
\arguments{
\item{height}{Body height in meters.}
}
\value{
A tibble with lower weight limits for each body mass index category.
}
\description{
Calculate the lower weight limits of body mass index categories based on body height in meters.
}
\examples{
bmi_limits(1.65)
}
