% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal2j.R
\name{cal2j}
\alias{cal2j}
\title{Convert Calories to Joules}
\usage{
cal2j(calories)
}
\arguments{
\item{calories}{A number of calories.}
}
\value{
A number of Joules.
}
\description{
Converts thermochemical calories to Joules.
}
\examples{
cal2j(11*10^6)
}
