% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulate_recipe.R
\name{emulate_recipe}
\alias{emulate_recipe}
\title{Emulate Recipe}
\usage{
emulate_recipe(recipe_ingredient_list, global_ingredient_list)
}
\arguments{
\item{recipe_ingredient_list}{A list of ingredients to include in the recipe
in the format of a list of key-value pairs (e.g. \code{c('sfa1' = 150, 'sfa17' = 37)}) and the number of grams of each ingredient in the recipe.}

\item{global_ingredient_list}{A list of all available ingredients imported by
\code{get_database_ingredients()}.}
}
\value{
A list of nutritional values by ingredient and a row of totals at the
end.
}
\description{
Gives the user the ability to calculate the nutrtional values of a recipe
without writing it to the database. Accepts a list of key-value pairs
specifiying the number of grams of a \code{food_number} from a database imported
by \code{get_database_ingredients()} (e.g. \code{c('sfa1' = 150, 'sfa17' = 37)}) as
input and outputs a list of nutritional values per ingredients and a row of
totals. A list of ingredients imported from the database must also be
provided.
}
\seealso{
\code{\link[=get_database_ingredients]{get_database_ingredients()}}
}
