% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_database.R
\docType{data}
\name{food_database}
\alias{food_database}
\title{Swedish Food Agency’s Food Database}
\format{
\subsection{food_database A data frame with 2,333 rows and 61 columns:}{

\describe{
\item{food_name}{Food name in Swedish.}
\item{food_number}{Database food number.}
\item{grouping}{Group to which the food belongs.}
\item{energy_kcal}{Energy contents in kilocalories}
\item{energy_kj}{Energy contents in kilojoules.}
\item{fat_total_g}{Total fat contents in grams.}
\item{protein_g}{Protein contents in grams.}
\item{carbohydrates_available_g}{Available carbohydrates in grams.}
\item{fibers_g}{Fiber contents in grams.}
\item{water_g}{Water contents in grams.}
\item{alcohol_g}{Alcohol contents in grams.}
\item{ash_g}{Ash contents in grams.}
\item{sugars_total_g}{Total sugar contents in grams.}
\item{monosaccharides_g}{Monosaccharide contents in grams.}
\item{disaccharides_g}{Disaccharide contents in grams.}
\item{added_sugar_g}{Added sugar contents in grams.}
\item{free_sugar_g}{Free sugar contents in grams.}
\item{whole_grain_total_g}{Total whole grain contents in grams.}
\item{sum_saturated_fatty_acids_g}{The sum of saturated fatty acid contents in grams.}
\item{fatty_acid_4_0_to_10_0_g}{The contents of butyric acid, caproic acid, and caprylic acid in grams.}
\item{lauric_acid_c12_0_g}{The contents of lauric acid in grams.}
\item{myristic_acid_c14_0_g}{The contents of myristic acid in grams.}
\item{palmitic_acid_c16_0_g}{The contents of palmitic acid in grams.}
\item{stearic_acid_c18_0}{The contents of stearic acid in grams.}
\item{arachidic_acid_c20_0_g}{The contents of arachidic acid in grams.}
\item{sum_monounsaturated_fatty_acids_g}{The sum of monosaturated fatty accids in grams.}
\item{palmitoleic_acid_c16_1_g}{The contents of palmitoleic acid in grams.}
\item{oleic_acid_c18_1_g}{The contents of oleic acid in grams.}
\item{sum_polyunsaturated_fatty_acids_g}{The sum of polyunsaturated fatty acid contents in grams.}
\item{linoleic_acid_c18_2_g}{The contents of linoleic acid in grams.}
\item{linolenic_acid_c18_3_g}{The contents of linolenic acid in grams.}
\item{arachidonic_acid_c20_4_g}{The contents of arachidonic acid in grams.}
\item{epa_c20_5_g}{The contents of eicosapentaenoic acid in grams.}
\item{dpa_c22_5_g}{The contents of docosapentaenoic acid in grams.}
\item{dha_c22_6_g}{The contents of docosahexaenoic acid in grams.}
\item{cholesterol_mg}{The cholesterol contents in milligrams.}
\item{vitamin_a_re_per_mcg}{The contents of vitamin A in retinol equivalents per microgram.}
\item{retinol_mcg}{The retinol contents in micrograms.}
\item{beta_carotene_mcg}{The beta carotene contents in micrograms.}
\item{vitamin_d_mcg}{The contents of vitamin D in micrograms.}
\item{vitamin_e_mg}{The contents of vitamin E in milligrams.}
\item{vitamin_k_mcg}{The contents of vitamin K in micrograms.}
\item{thiamine_mg}{The thiamine contents in milligrams.}
\item{riboflavin_mg}{The riboflavin contents in milligrams.}
\item{niacin_mg}{The niacin contents in milligrams.}
\item{niacin_equivalents_ne_per_mg}{Niacin equivalents per milligram.}
\item{vitamin_b6_mg}{The contents of vitamin B6 in milligrams.}
\item{folate_mcg}{The folate contents in micrograms.}
\item{vitamin_b12_mcg}{The contents of vitamin B12 in micrograms.}
\item{vitamin_c_mg}{The contents of vitamin C in milligrams.}
\item{phosphorous_mg}{The phosphorous contents in milligrams.}
\item{iodine_mcg}{The iodine contents in micrograms.}
\item{iron_mg}{The iron contents in milligrams.}
\item{calcium_mg}{The calcium contents in milligrams.}
\item{potassium_mg}{The potassium contents in milligrams.}
\item{magnesium_mg}{The magnesium contents in milligrams.}
\item{sodium_mg}{The sodium contents in milligrams.}
\item{sodium_chloride_g}{The sodium chloride contens in milligrams.}
\item{selenium_mcg}{The selenium contens in micrograms.}
\item{zink_mg}{The zink contents in milligrams.}
\item{waste_pct}{The proportion of the food comprised by waste products, expressed in percent.}
}
}
}
\source{
\url{https://www.livsmedelsverket.se/livsmedelsdatabasen}
}
\usage{
food_database
}
\description{
The data in the dataset prepared below is derived from the Swedish language
version of \href{https://www.livsmedelsverket.se/livsmedelsdatabasen}{the Swedish Food Agency food database}, version
2022-05-24, available under a \href{http://www.creativecommons.se/wp-content/uploads/2015/01/CreativeCommons-Erkännande-4.0.pdf}{Creative Commons Attribution 4.0 license}.
}
\keyword{datasets}
