% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_ingredients.R
\name{get_database_ingredients}
\alias{get_database_ingredients}
\title{Get Database Ingredients}
\usage{
get_database_ingredients(database_path)
}
\arguments{
\item{database_path}{File path to database.}
}
\value{
A tibble of nutrition facts.
}
\description{
Imports nutrition facts for ingredients in a database formatted \emph{ad modum}
\href{https://github.com/JonasEngstrom/food_database_scripts}{food_database_scripts}.
}
