% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_recipes.R
\name{get_database_recipes}
\alias{get_database_recipes}
\title{Get Database Recipes}
\usage{
get_database_recipes(database_path)
}
\arguments{
\item{database_path}{File path to database.}
}
\value{
A tibble of nutrition facts.
}
\description{
Imports nutrition facts for recipes in a database formatted ad modum
\href{https://github.com/JonasEngstrom/food_database_scripts}{food_database_scripts}.
}
