% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibw.R
\name{ibw}
\alias{ibw}
\title{Calculate Ideal Body Weight}
\usage{
ibw(height, female = TRUE, formula = "Broca")
}
\arguments{
\item{height}{Height in meters.}

\item{female}{Boolean indicating female sex.}

\item{formula}{The formula to use for calculations, se documentation for available options.}
}
\value{
A value in kilograms.
}
\description{
Calculate the ideal body weight based on height and sex, using one of the following formulas:
}
\details{
\itemize{
\item Broca, 1871
\item Hamwi, 1964 (without consideration for frame size)
\item Devine, 1974
\item Robinson, 1983
\item Miller, 1983
\item Hammond, 2000
\item Lemmens, 2005
\item Deitel, 2003 (without consideration for frame size)
}
}
\examples{
ibw(1.65)
}
