% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in2m.R
\name{in2m}
\alias{in2m}
\title{Convert Inches to Meters}
\usage{
in2m(inches)
}
\arguments{
\item{inches}{A number of inches.}
}
\value{
A number of meters.
}
\description{
Converts inches to meters as per the international yard defined in 1959.
}
\examples{
in2m(12)
}
