% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j2cal.R
\name{j2cal}
\alias{j2cal}
\title{Convert Joules to Calories}
\usage{
j2cal(joules)
}
\arguments{
\item{joules}{A number of Joules.}
}
\value{
A number of calories.
}
\description{
Converts Joules to thermochemical calories.
}
\examples{
j2cal(11*10^6)
}
