% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg2lb.R
\name{kg2lb}
\alias{kg2lb}
\title{Convert Kilograms to Pounds}
\usage{
kg2lb(kilograms)
}
\arguments{
\item{kilograms}{A number of kilograms.}
}
\value{
A number of pounds.
}
\description{
Converts a value in kilograms to an exact value of international avoirdupois
pounds as defined in 1959.
}
\examples{
kg2lb(71)
}
