% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb2kg.R
\name{lb2kg}
\alias{lb2kg}
\title{Convert Pounds to Kilograms}
\usage{
lb2kg(pounds)
}
\arguments{
\item{pounds}{A number of pounds.}
}
\value{
A number of kilograms.
}
\description{
Converts a value in international avoirdupois pounds as defined in 1959 to an exact value of kilograms.
}
\examples{
lb2kg(156)
}
