% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbw.R
\name{lbw}
\alias{lbw}
\title{Calculate Lean Body Weight}
\usage{
lbw(height, weight, female = TRUE)
}
\arguments{
\item{height}{Height in meters.}

\item{weight}{Weight in kilograms.}

\item{female}{Boolean indicating female sex.}
}
\value{
A number of kilograms.
}
\description{
Calculates lean body weight in kilograms according to Janmahasatian.
}
\examples{
lbw(1.81, 71.2, female = FALSE)
}
