% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2in.R
\name{m2in}
\alias{m2in}
\title{Convert Meters to Inches}
\usage{
m2in(meters)
}
\arguments{
\item{meters}{A number of meters.}
}
\value{
A number of inches.
}
\description{
Converts meters to inches as per the international yard defined in 1959.
}
\examples{
m2in(12)
}
