% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight_over_time.R
\name{plot_weight_over_time}
\alias{plot_weight_over_time}
\title{Plot Weight Over Time}
\usage{
plot_weight_over_time(weight_tibble, height)
}
\arguments{
\item{weight_tibble}{A tibble with two columns named \emph{date} and \emph{weight}, the former formatted as \code{POSIXct} and the latter as \code{double}.}

\item{height}{A body height in meters.}
}
\value{
A ggplot2.
}
\description{
Plots weight over time on a background divided \href{https://www.who.int/europe/news-room/fact-sheets/item/a-healthy-lifestyle---who-recommendations}{according to World Health Organization body mass index classes}.
}
\examples{
tibble::tibble(
  date=c(
    as.POSIXct('2016-01-04'),
    as.POSIXct('2020-06-07')
  ),
  weight = c(
    75,
    72
  )
 ) |> plot_weight_over_time(1.81)
}
